Placed getContainer() and getConfig() in a separate trait to make explicit usage possible
